pyproject.toml
[tool.poetry]
name = "project"
version = "0.1.0"
description = "This is a sample project"
authors = ["pk karki <prabin@abhiyantrik.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.1.1"
psycopg2 = "^2.8.6"
slackclient = "^2.8.2"
requests = "^2.24.0"
djangorestframework = "^3.11.1"
bandit = "^1.6.2"
pre-commit = "^2.7.1"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
flake8 = "^3.8.3"
mypy = "^0.782"
isort = "^5.5.2"
pre-commit = "^2.7.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
